Prevent unsafe narrowing: base/, random trybot redness edition
This fixes things that the trybots turned up in my initial tests at
turning on the narrowing config for base.

Bug: 1292951
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3769281
Owners-Override: danakj <danakj@chromium.org>
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Mike Wittman <wittman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1026009}

+++ base/allocator/winheap_stubs_win.cc
@@ -18,6 +18,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "base/bits.h"
 #include "base/check_op.h"
+#include "base/numerics/safe_conversions.h"
 
 namespace base {
 namespace allocator {
@@ -123,7 +124,7 @@ void* AlignAllocation(void* ptr, size_t alignment) {
   // Write the prefix.
   AlignedPrefix* prefix = reinterpret_cast<AlignedPrefix*>(address) - 1;
   prefix->original_allocation_offset =
-      address - reinterpret_cast<uintptr_t>(ptr);
+      checked_cast<unsigned int>(address - reinterpret_cast<uintptr_t>(ptr));
 #if DCHECK_IS_ON()
   prefix->magic = AlignedPrefix::kMagic;
 #endif  // DCHECK_IS_ON()
