+++ base/memory/page_size_nacl.cc
@@ -11,7 +11,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 namespace base {
 
 size_t GetPageSize() {
-  return getpagesize();
+  return static_cast<size_t>(getpagesize());
 }
 
 }  // namespace base
